+++ src/main/java/seb/Event.java
@@ -1,4 +1,4 @@
-package duke;
+package seb;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
