+++ src/main/java/seb/Launcher.java
@@ -1,0 +1,9 @@
+package seb;
+
+import javafx.application.Application;
+
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
