+++ src/main/java/seb/Parser.java
@@ -1,7 +1,7 @@
-package duke;
+package seb;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import duke.command.*;
+import seb.command.*;
 
 public class Parser {
     // task must not be empty, can contain spaces, must not start with space, time/date can be empty
