+++ src/main/java/seb/Storage.java
@@ -1,4 +1,4 @@
-package duke;
+package seb;
 import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
