+++ src/main/java/seb/TimeParser.java
@@ -1,4 +1,4 @@
-package duke;
+package seb;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
