+++ src/main/java/seb/Ui.java
@@ -1,4 +1,4 @@
-package duke;
+package seb;
 import java.util.Scanner;
 
 /**
