+++ src/main/java/seb/UnknownInputException.java
@@ -1,4 +1,4 @@
-package duke;
+package seb;
 public class UnknownInputException extends Exception {
     public UnknownInputException() {
         super("     OOPS!!! I'm just a chatbot. I can only handle task schedule right now.");
