+++ src/main/java/seb/command/Command.java
@@ -1,5 +1,5 @@
-package duke.command;
-import duke.*;
+package seb.command;
+import seb.*;
 public interface Command {
     String execute(TaskList tasks, Storage storage);
 }
