+++ src/main/java/seb/command/ListCommand.java
@@ -1,5 +1,5 @@
-package duke.command;
-import duke.*;
+package seb.command;
+import seb.*;
 public class ListCommand implements Command {
     @Override
     public String execute(TaskList tasks, Storage storage) {
